requirements.txt
streamlit
pytesseract
edge-tts
pymupdf
Pillow